% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathway_score_seurat.R
\name{aggregate_pathway_score}
\alias{aggregate_pathway_score}
\title{transform pathway score per cell to pathway score per cell type}
\usage{
aggregate_pathway_score(
  object,
  cluster = "cell_type",
  avg_method = c("mean", "median")
)
}
\arguments{
\item{object}{Seurat object from sc.pathway.score.seurat}

\item{cluster}{the feature to aggregate}

\item{avg_method}{aggregate method, mean or median}
}
\value{
a data frame
}
\description{
transform pathway score per cell to pathway score per cell type
}
