% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_markers.R
\name{convert_symbol}
\alias{convert_symbol}
\title{Covert between mouse and human symbols}
\usage{
convert_symbol(genesig = NULL, to = "mouse")
}
\arguments{
\item{genesig}{vector, gene signature}

\item{to}{'human' or 'mouse'}
}
\description{
Covert between mouse and human symbols
}
