% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichpathway.R
\name{dotplot_encich_clusterprofiler}
\alias{dotplot_encich_clusterprofiler}
\title{use clusterprofiler to do enrichment analysis per cell type}
\usage{
dotplot_encich_clusterprofiler(object, fun, ...)
}
\arguments{
\item{object}{Seurat object}

\item{fun}{fun param in clusterProfiler::compareCluster}
}
\description{
use clusterprofiler to do enrichment analysis per cell type
}
