% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathway_score_seurat.R
\name{dotplot_pathway_score}
\alias{dotplot_pathway_score}
\title{dotplot to visualize pathway score per cell type}
\usage{
dotplot_pathway_score(df, topn = 5, normalization = c("zeroone", "z-score"))
}
\arguments{
\item{df}{data frame from function `aggregate_pathway_score`}
}
\value{
`ggplot2` object
}
\description{
dotplot to visualize pathway score per cell type
}
