% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{FeaturePlot_gene_pos}
\alias{FeaturePlot_gene_pos}
\title{Function to plot postive cells or negative cells with two colors}
\usage{
FeaturePlot_gene_pos(
  object,
  gene,
  cutoff = 0,
  color_value = c("grey", "red"),
  legend = FALSE,
  reduction = "umap"
)
}
\arguments{
\item{object}{Seurat object}

\item{gene}{gene name}

\item{cutoff}{cutoff}

\item{color_value}{color value  c("grey","red")}

\item{legend}{whether use legend, default: FALSE}

\item{reduction}{what is the reduction, default: umap}
}
\value{
a ggplot2 object
}
\description{
Function to plot postive cells or negative cells with two colors
}
