% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon_tfs_utils.R
\name{get_pseudocell}
\alias{get_pseudocell}
\title{get pseudocell for `Seurat` object}
\usage{
get_pseudocell(
  object = NULL,
  slot = "data",
  assay = "RNA",
  group_id = "cell_type",
  seed = 123,
  pseudocell_size = 5
)
}
\arguments{
\item{object}{`Seurat` object}

\item{group_id}{character, the column to aggregate in meta data}

\item{seed}{set.seed}

\item{pseudocell_size}{pseudocell size (default: 5)}
}
\description{
get pseudocell for `Seurat` object
}
