% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_markers.R
\name{write.gmt}
\alias{write.gmt}
\title{Convert list to gmt file}
\usage{
write.gmt(geneSet = NULL, gmt_file = NULL)
}
\arguments{
\item{geneSet}{a list object}

\item{gmt_file}{output file name .gmt}
}
\description{
Convert list to gmt file
}
