% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon_tfs_utils.R
\name{write_pyscenic_mtx}
\alias{write_pyscenic_mtx}
\title{write mtx for first step in pyscenic}
\usage{
write_pyscenic_mtx(
  object,
  path = "/export/bioinfo-team/home/xiongy/pyscenic/mtx.tsv",
  slot = "data",
  assay = "RNA"
)
}
\arguments{
\item{object}{`Seurat` object}

\item{path}{write path}

\item{slot}{slot for `Seurat` object}

\item{assay}{assay for `Seurat` object}
}
\description{
write mtx for first step in pyscenic
}
